moved design classes from :app
