+++ app/src/main/java/srimani7/apps/feedfly/MainActivity.kt
@@ -12,9 +12,9 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.core.view.WindowCompat
+import srimani7.apps.feedfly.core.design.TheSecretDairyTheme
 import srimani7.apps.feedfly.data.AppTheme
 import srimani7.apps.feedfly.navigation.URL_REGEX
-import srimani7.apps.feedfly.ui.theme.TheSecretDairyTheme
 import srimani7.apps.feedfly.viewmodel.HomeViewModal
 
 class MainActivity : ComponentActivity() {
