+++ core/design/src/main/java/srimani7/apps/feedfly/core/design/Theme.kt
@@ -1,4 +1,4 @@
-package srimani7.apps.feedfly.ui.theme
+package srimani7.apps.feedfly.core.design
 
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
@@ -9,9 +9,7 @@ import androidx.compose.material3.dynamicLightColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
-import com.google.accompanist.systemuicontroller.rememberSystemUiController
 
 private val LightColorScheme = lightColorScheme(
     primary = md_theme_light_primary,
@@ -83,7 +81,6 @@ fun TheSecretDairyTheme(
     dynamicColor: Boolean = true,
     content: @Composable () -> Unit
 ) {
-    val systemUiController = rememberSystemUiController()
 
     val colorScheme = when {
         dynamicColor && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
@@ -99,11 +96,7 @@ fun TheSecretDairyTheme(
         colorScheme = colorScheme,
         typography = Typography,
     ) {
-        DisposableEffect(systemUiController, darkTheme) {
-            systemUiController.setSystemBarsColor(
-                color = Color.Transparent,
-                darkIcons = !darkTheme
-            )
+        DisposableEffect(darkTheme) {
             onDispose {}
         }
         content()
